Add boolean flag for block instances.
Rhino always adds mesh data in world space,
except for block instances. This effectively
means that transforms should never be applied.

If a scene contains only one block instance
the original code would apply transform, which
after manipulating the scene in an interactive
session results in geometry snapping to
world origo. This ensures we never apply transform
when not necessary.