Add a new way to control render loop.
Instead of having a new thread being issued for the original
run() function we break it up in a set of new functions.
* prepare_run()
  set up the session for rendering
* sample()
  call as many times as needed to reach the requested amount of
  samples. When the requested amount has been reached a call to
  sample() will return false (i.e. nothing more to render). One call
  is equivalent to rendering one sample of the scene (counted from
  camera rays)
* end_run()
  clean up.

Together these three functions do what the original run() function
did, but there is more control over what happens when, eliminating
the need for locks all around.