Support shader on object with shader indirection.
For Rhino instances it is more useful to have the
shader information on the object. It is very unlikely
that one mesh will have more than one shader applied.

In Rhino an object that gets assigned multipe materials
via sub-object selection eventually ends up in the
changequeue as separate mesh instances, and thus will
not have multiple shaders.

If the need arises after all we can always make the member
into an array of Shader pointers and adapt rendering code
accordingly.

Interesting note: we could get rid of the vector of
per-face shaders, potentially saving a lot of memory.